requirements.txt
fastapi==0.115.7
mangum==0.17.0
openai==1.60.2
pydantic==2.10.6
python-dotenv==1.0.1
